requirements.txt
fastapi
uvicorn
spacy
pandas
python-multipart
jinja2
python-docx
pdfminer.six
